+++ Chapter_5/Lesson-5.7.1/src/main/java/ru/apetrov/OrderBook/ParserStAX.java
@@ -54,7 +54,7 @@ public class ParserStAX {
                         String bookName = parser.getAttributeValue(0);
                         String operation = parser.getAttributeValue(1);
                         double price = Double.parseDouble(parser.getAttributeValue(2));
-                        double volume = Double.parseDouble(parser.getAttributeValue(3));
+                        int volume = Integer.parseInt(parser.getAttributeValue(3));
                         int orderId = Integer.parseInt(parser.getAttributeValue(4));
                         add(new Order(bookName, operation, price, volume, orderId));
